+++ NEWS.md
@@ -1,6 +1,7 @@
 # spant 0.13.0
 * Added ylim and y_scale options to fit plotting.
 * Added %$% operator from magrittr package.
+* Added an interpolation option to calc_spec_snr.
 
 # spant 0.12.0
 * Fit results stackplot now has the option to display labels.