+++ NEWS.md
@@ -1,3 +1,5 @@
+# spant 0.14.0
+
 # spant 0.13.0
 * Various bug fixes for Siemens TWIX reader.
 * rats and tdsr functions now use the mean spectrum as the default reference.